Add tests for run checks (#97)
* Add tests to checks if all missing global attributes are detected

* Add gitpython to requirements

* Update test_run_checks.py after rebase and use tmpdir for demo data and checker output

* Refactor file_checks dict into JSON which is located in demo_data

* Check if verbose file exits before removing it

* Update pytest.yml and CHANGELOG.md

+++ run_checks.py
@@ -122,7 +122,7 @@ def cmd_string_checker(io_in, idiryml_in):
     ofiles_checker_string = " ".join(io_in[1])
 
     # Output string creation
-    return 'compliance-checker --y ' + idiryml_in + '/atmodat_standard_checks.yml -f json_new -f text ' + \
+    return 'compliance-checker --y ' + idiryml_in + 'atmodat_standard_checks.yml -f json_new -f text ' + \
            ofiles_checker_string + ' --test atmodat_standard:3.0 ' + ifile_in_string
 
 
@@ -215,7 +215,8 @@ def run_checks(ifile_in, verbose_in, check_types_in, cfversion_in, opath_file, i
                         print(f_verbose.read())
             # Clean-up
             if check == 'atmodat':
-                os.remove(file_verbose)
+                if os.path.isfile(file_verbose):
+                    os.remove(file_verbose)
     return
